requirements.txt
scikit-learn>=1.0.2
pandas>=1.3.0
numpy>=1.21.0
streamlit>=1.10.0
altair>=4.2.0
nltk>=3.6.0
python-docx>=0.8.11
PyPDF2>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0 